+++ src/main/java/frc/robot/RobotContainer.java
@@ -19,10 +19,10 @@ import edu.wpi.first.wpilibj.DoubleSolenoid;
 import edu.wpi.first.wpilibj.SerialPort.Port;
 import edu.wpi.first.wpilibj2.command.Command;
 import edu.wpi.first.wpilibj2.command.Commands;
-import frc.robot.Subsystem.Gyro.NavX;
-import frc.robot.Subsystem.Swerve.DoubleSolenoidSwerveShifter;
-import frc.robot.Subsystem.Swerve.ShiftingSwerveDrive;
-import frc.robot.Subsystem.Swerve.ShiftingSwerveModule;
+import frc.robot.Subsystems.Gyro.NavX;
+import frc.robot.Subsystems.Swerve.DoubleSolenoidSwerveShifter;
+import frc.robot.Subsystems.Swerve.ShiftingSwerveDrive;
+import frc.robot.Subsystems.Swerve.ShiftingSwerveModule;
 import frc.robot.util.ConfigUtils;
 import frc.robot.util.SwerveUtils;
 import lib.Config.DoubleSolenoidSwerveShifterConfig;
