+++ src/main/java/frc/robot/Subsystems/Gyro/NavX.java
@@ -2,7 +2,7 @@
 // Open Source Software; you can modify and/or share it under the terms of
 // the WPILib BSD license file in the root directory of this project.
 
-package frc.robot.Subsystem.Gyro;
+package frc.robot.Subsystems.Gyro;
 
 import com.kauailabs.navx.frc.AHRS;
