+++ src/main/java/frc/robot/Subsystems/Swerve/ShiftingSwerveModule.java
@@ -2,7 +2,7 @@
 // Open Source Software; you can modify and/or share it under the terms of
 // the WPILib BSD license file in the root directory of this project.
 
-package frc.robot.Subsystem.Swerve;
+package frc.robot.Subsystems.Swerve;
 
 import java.util.concurrent.atomic.AtomicInteger;
